refactor(gallery): split db service
split db service to users and images

MODULE3-PART2-dynamodb
